feat: create EmployeeProjectsCellContainer to handle data fetching for employee projects
